+++ src/common/utils.spec.ts
@@ -1,7 +1,7 @@
 import { DeepMocked } from '@golevelup/ts-jest';
 import { Request } from 'express';
-import { backLink } from './utils';
-import { createMockRequest } from './create-mock-request';
+import { backLink, formatDate } from './utils';
+import { createMockRequest } from '../testutils/create-mock-request';
 
 describe('utils', () => {
   describe('backLink', () => {
@@ -46,6 +46,13 @@ describe('utils', () => {
       it('returns undefined', () => {
         expect(backLink(req)).toBeUndefined();
       });
+    });
+  });
+
+  describe('formatDate', () => {
+    it('returns a date in DD-MM-YYYY format', () => {
+      // Months are zero-indexed
+      expect(formatDate(new Date(1999, 5, 23))).toEqual('23-06-1999');
     });
   });
 });