+++ src/professions/admin/professions.presenter.ts
@@ -8,29 +8,47 @@ import { NationsCheckboxPresenter } from '../../nations/nations-checkbox.present
 import { Profession } from '../profession.entity';
 import { FilterInput } from '../interfaces/filter-input.interface';
 import { IndexTemplate } from './interfaces/index-template.interface';
-import { ProfessionSearchResultPresenter } from './profession-search-result.presenter';
+import { ListEntryPresenter } from './list-entry.presenter';
 import { ProfessionsSorter } from '../helpers/professions-sorter';
+import { Organisation } from '../../organisations/organisation.entity';
+import { OrganisationsCheckboxPresenter } from '../../organisations/organisations-checkbox-presenter';
 
-export class SearchPresenter {
+export type ProfessionsPresenterView = 'overview' | 'single-organisation';
+
+export class ProfessionsPresenter {
   constructor(
     private readonly filterInput: FilterInput,
+    private readonly userOrganisaiton: Organisation | null,
     private readonly allNations: Nation[],
+    private readonly allOrganisations: Organisation[],
     private readonly allIndustries: Industry[],
     private readonly filteredProfessions: Profession[],
-    private readonly i18nService: I18nService,
     private readonly request: Request,
+    private readonly i18nService: I18nService,
   ) {}
 
-  async present(): Promise<IndexTemplate> {
+  async present(view: ProfessionsPresenterView): Promise<IndexTemplate> {
+    const organisation =
+      view === 'overview'
+        ? await this.i18nService.translate('app.ukcpq')
+        : this.userOrganisaiton.name;
+
+    const headings = await ListEntryPresenter.headings(this.i18nService, view);
+
     const nationsCheckboxArgs = await new NationsCheckboxPresenter(
       this.allNations,
-      this.filterInput.nations,
+      this.filterInput.nations || [],
       this.i18nService,
+    ).checkboxArgs();
+
+    const organisationsCheckboxArgs = await new OrganisationsCheckboxPresenter(
+      this.allOrganisations,
+      this.filterInput.organisations || [],
     ).checkboxArgs();
 
     const industriesCheckboxArgs = await new IndustriesCheckboxPresenter(
       this.allIndustries,
-      this.filterInput.industries,
+      this.filterInput.industries || [],
       this.i18nService,
     ).checkboxArgs();
 
@@ -40,23 +58,29 @@ export class SearchPresenter {
 
     const displayProfessions = await Promise.all(
       sortedProfessions.map(async (profession) =>
-        new ProfessionSearchResultPresenter(
-          profession,
-          this.i18nService,
-        ).present(),
+        new ListEntryPresenter(profession, this.i18nService).tableRow(view),
       ),
     );
 
     return {
+      view,
+      organisation,
+      headings,
       professions: displayProfessions,
       nationsCheckboxArgs,
+      organisationsCheckboxArgs,
       industriesCheckboxArgs,
+      changedByCheckboxArgs: [],
       filters: {
-        keywords: this.filterInput.keywords,
-        nations: this.filterInput.nations.map((nation) => nation.name),
-        industries: this.filterInput.industries.map(
+        keywords: this.filterInput.keywords || '',
+        nations: (this.filterInput.nations || []).map((nation) => nation.name),
+        organisations: (this.filterInput.organisations || []).map(
+          (organisation) => organisation.name,
+        ),
+        industries: (this.filterInput.industries || []).map(
           (industry) => industry.name,
         ),
+        changedBy: [],
       },
       backLink: backLink(this.request),
     };