+++ src/professions/helpers/filter.helper.ts
@@ -1,5 +1,6 @@
-import { Industry } from '../../../industries/industry.entity';
-import { Profession } from '../../../professions/profession.entity';
+import { Industry } from '../../industries/industry.entity';
+import { Organisation } from '../../organisations/organisation.entity';
+import { Profession } from '../../professions/profession.entity';
 import { FilterInput } from '../interfaces/filter-input.interface';
 
 export class FilterHelper {
@@ -13,9 +14,14 @@ export class FilterHelper {
       unfilteredProfessions,
     );
 
-    const industryFilteredProfessions = this.filterByIndustry(
+    const organisationFilteredProfessions = this.filterByOrganisation(
       filterInput,
       nationFilteredProfessions,
+    );
+
+    const industryFilteredProfessions = this.filterByIndustry(
+      filterInput,
+      organisationFilteredProfessions,
     );
 
     const keywordFilteredProfessions = this.filterByKeyword(
@@ -30,9 +36,7 @@ export class FilterHelper {
     filterInput: FilterInput,
     professions: Profession[],
   ): Profession[] {
-    if (filterInput.nations.length === 0) {
-      return professions;
-    } else {
+    if (filterInput.nations?.length) {
       const filterNationCodes = filterInput.nations.map(
         (nation) => nation.code,
       );
@@ -40,6 +44,24 @@ export class FilterHelper {
       return professions.filter((profession) =>
         this.isNationOverlap(profession.occupationLocations, filterNationCodes),
       );
+    } else {
+      return professions;
+    }
+  }
+
+  private filterByOrganisation(
+    filterInput: FilterInput,
+    professions: Profession[],
+  ): Profession[] {
+    if (filterInput.organisations?.length) {
+      return professions.filter((profession) =>
+        this.isOrganisationOverlap(
+          profession.organisation ? [profession.organisation] : [],
+          filterInput.organisations,
+        ),
+      );
+    } else {
+      return professions;
     }
   }
 
@@ -47,12 +69,12 @@ export class FilterHelper {
     filterInput: FilterInput,
     professions: Profession[],
   ): Profession[] {
-    if (filterInput.industries.length == 0) {
-      return professions;
-    } else {
+    if (filterInput.industries?.length) {
       return professions.filter((profession) =>
         this.isIndustryOverlap(profession.industries, filterInput.industries),
       );
+    } else {
+      return professions;
     }
   }
 
@@ -62,12 +84,12 @@ export class FilterHelper {
   ): Profession[] {
     const searchTerms = this.getSearchTerms(filterInput);
 
-    if (searchTerms.length === 0) {
-      return professions;
-    } else {
+    if (searchTerms?.length) {
       return professions.filter((profession) =>
         this.matchesKeywords(profession, searchTerms),
       );
+    } else {
+      return professions;
     }
   }
 
@@ -78,17 +100,28 @@ export class FilterHelper {
     return nationCodes1.some((code) => nationCodes2.includes(code));
   }
 
+  private isOrganisationOverlap(
+    organisations1: Organisation[],
+    organisations2: Organisation[],
+  ): boolean {
+    return organisations1.some((organisation1) => {
+      return organisations2.some(
+        (organisation2) => organisation1.id === organisation2.id,
+      );
+    });
+  }
+
   private isIndustryOverlap(
     industries1: Industry[],
     industries2: Industry[],
   ): boolean {
     return industries1.some((industry1) => {
-      return industries2.some((industry2) => industry1.id == industry2.id);
+      return industries2.some((industry2) => industry1.id === industry2.id);
     });
   }
 
   private getSearchTerms(filterInput: FilterInput): string[] {
-    return filterInput.keywords
+    return (filterInput.keywords || '')
       .toLowerCase()
       .split(' ')
       .filter((term) => term !== '');