+++ src/professions/search/interfaces/filter-input.interface.ts
@@ -1,8 +1,0 @@
-import { Industry } from '../../../industries/industry.entity';
-import { Nation } from '../../../nations/nation';
-
-export interface FilterInput {
-  nations: Nation[];
-  industries: Industry[];
-  keywords: string;
-}